+++ packages/core/src/parse/tokens/border.ts
@@ -1,4 +1,4 @@
-import type {ParsedBorderToken} from '../../token.js';
+import type {BorderToken, ParsedBorderToken} from '../../token.js';
 import {isObj} from '../../util.js';
 import {normalizeColorValue} from './color.js';
 import {normalizeDimensionValue} from './dimension.js';
@@ -19,27 +19,13 @@ import {normalizeStrokeStyleValue} from './stroke-style.js';
 
 export function normalizeBorderValue(value: unknown): ParsedBorderToken['$value'] {
   if (!isObj(value)) throw new Error(`Expected object, received ${Array.isArray(value) ? 'array' : typeof value}`);
-  const normalized = {} as ParsedBorderToken['$value'];
-  if (!Object.keys(value as any).length) throw new Error(`Must specify at least 1 property for border`);
-  for (const k of Object.keys(value as ParsedBorderToken['$value'])) {
-    // note: all properties are required
-    switch (k) {
-      case 'color': {
-        normalized.color = normalizeColorValue((value as any).color);
-        break;
-      }
-      case 'width': {
-        normalized.width = normalizeDimensionValue((value as any).width);
-        break;
-      }
-      case 'style': {
-        normalized.style = normalizeStrokeStyleValue((value as any).style);
-        break;
-      }
-      default: {
-        throw new Error(`Unknown property "${k}" on border token`);
-      }
-    }
-  }
-  return normalized;
+  const tokenValue = value as BorderToken['$value'];
+  if (!('color' in tokenValue)) throw new Error(`Token missing required "color" property`);
+  if (!('width' in tokenValue)) throw new Error(`Token missing required "width" property`);
+  if (!('style' in tokenValue)) throw new Error(`Token missing required "style" property`);
+  return {
+    color: normalizeColorValue(tokenValue.color),
+    width: normalizeDimensionValue(tokenValue.width),
+    style: normalizeStrokeStyleValue(tokenValue.style),
+  };
 }